Replace container to dry-system and use one container in all application
