+++ config/initializers/application.rb
@@ -1,1 +1,1 @@
-OSSBoard::Application.finalize!
+Container.finalize!
