--- lib/ossboard.rb
@@ -1,2 +1,0 @@
-module OSSBoard
-end
