+++ spec/ossboard/repositories/bloked_user_repository_spec.rb
@@ -1,5 +1,5 @@
 RSpec.describe BlokedUserRepository do
-  before { OSSBoard::Application[:redis].with(&:flushdb) }
+  before { Container[:redis].with(&:flushdb) }
 
   let(:repo) { BlokedUserRepository.new }
