+++ spec/spec_helper.rb
@@ -22,8 +22,8 @@ require 'fabrication'
 
 require_relative '../config/environment'
 require 'dry/container/stub'
-OSSBoard::Application.enable_stubs!
-OSSBoard::Application.stub(:redis, ConnectionPool.new(size: 10, timeout: 3) { MockRedis.new })
+Container.enable_stubs!
+Container.stub(:redis, ConnectionPool.new(size: 10, timeout: 3) { MockRedis.new })
 
 Hanami.boot
