+++ system/boot/twitter.rb
@@ -1,4 +1,4 @@
-OSSBoard::Application.finalize(:twitter) do |container|
+Container.finalize(:twitter) do |container|
   client = Twitter::REST::Client.new do |config|
     config.consumer_key        = ENV["TWITTER_CONSUMER_KEY"]
     config.consumer_secret     = ENV["TWITTER_CONSUMER_SECRET"]
