+++ system/import.rb
@@ -1,5 +1,3 @@
 require_relative './container'
 
-module OSSBoard
-  Import = Application.injector
-end
+Import = Container.injector
